test(e2e): split apt action runtime coverage
